bug 14975: Scanner needs to be faster, smaller and more-reliable for TinySC Make maintenance of TrackPersistence optional (controlled by --nostatistics).
Bug: 14975
git-svn-id: file:///home/awy/mirror/slim/7.5/branches/embedded/server@29095 62299810-d8cb-41fd-93b7-d32162e5a4a4

+++ Slim/Music/Import.pm
@@ -386,14 +386,16 @@ sub runScanPostProcessing {
 	
 	Slim::Music::Artwork->precacheAllArtwork;
 	
-	# Look for and import persistent data migrated from 7.3.3
-	my ($dir) = Slim::Utils::OSDetect::dirsFor('prefs');
-	my $json = catfile( $dir, 'tracks_persistent.json' );
-	if ( -e $json ) {
-		$log->error('Migrating persistent track information from MySQL');
-		
-		if ( Slim::Schema::TrackPersistent->import_json($json) ) {
-			unlink $json;
+	if (main::STATISTICS) {
+		# Look for and import persistent data migrated from MySQL
+		my ($dir) = Slim::Utils::OSDetect::dirsFor('prefs');
+		my $json = catfile( $dir, 'tracks_persistent.json' );
+		if ( -e $json ) {
+			$log->error('Migrating persistent track information from MySQL');
+			
+			if ( Slim::Schema::TrackPersistent->import_json($json) ) {
+				unlink $json;
+			}
 		}
 	}
